debugging

+++ transform/snowflake-dbt/models/gitlab_dotcom/base/gitlab_dotcom_users.sql
@@ -72,7 +72,7 @@ WITH source AS (
       group_view::INTEGER                                              AS group_views,
       managing_group_id::INTEGER                                       AS managing_group_id,
       bot_type::INTEGER                                                AS bot_type,
-      source.role                                                      AS role_id,
+      source.role::INTEGER                                             AS role_id,
       {{user_role_mapping(user_role='source.role')}}::VARCHAR          AS role
 
     FROM source
